api refactor - elrond-wasm-node reorg
