+++ build-wasm.sh
@@ -4,25 +4,25 @@
 
 ### BUILD ALL CONTRACTS ###
 
-erdpy --verbose contract build "contracts/benchmarks/str-repeat"
-erdpy --verbose contract build "contracts/examples/adder"
-erdpy --verbose contract build "contracts/examples/crowdfunding-egld"
-erdpy --verbose contract build "contracts/examples/crowdfunding-erc20"
-erdpy --verbose contract build "contracts/examples/crowdfunding-esdt"
-erdpy --verbose contract build "contracts/examples/crypto-bubbles"
-erdpy --verbose contract build "contracts/examples/factorial"
-erdpy --verbose contract build "contracts/examples/lottery-egld"
-erdpy --verbose contract build "contracts/examples/lottery-erc20"
-erdpy --verbose contract build "contracts/examples/lottery-esdt"
-erdpy --verbose contract build "contracts/examples/multisig"
-erdpy --verbose contract build "contracts/examples/non-fungible-tokens"
-erdpy --verbose contract build "contracts/examples/ping-pong-egld"
-erdpy --verbose contract build "contracts/examples/simple-erc20"
-erdpy --verbose contract build "contracts/feature-tests/abi-tester"
-erdpy --verbose contract build "contracts/feature-tests/basic-features"
-erdpy --verbose contract build "contracts/feature-tests/async/async-alice"
-erdpy --verbose contract build "contracts/feature-tests/async/async-bob"
-erdpy --verbose contract build "contracts/feature-tests/use-module"
+erdpy --verbose contract build "contracts/benchmarks/str-repeat" || return 1
+erdpy --verbose contract build "contracts/examples/adder" || return 1
+erdpy --verbose contract build "contracts/examples/crowdfunding-egld" || return 1
+erdpy --verbose contract build "contracts/examples/crowdfunding-erc20" || return 1
+erdpy --verbose contract build "contracts/examples/crowdfunding-esdt" || return 1
+erdpy --verbose contract build "contracts/examples/crypto-bubbles" || return 1
+erdpy --verbose contract build "contracts/examples/factorial" || return 1
+erdpy --verbose contract build "contracts/examples/lottery-egld" || return 1
+erdpy --verbose contract build "contracts/examples/lottery-erc20" || return 1
+erdpy --verbose contract build "contracts/examples/lottery-esdt" || return 1
+erdpy --verbose contract build "contracts/examples/multisig" || return 1
+erdpy --verbose contract build "contracts/examples/non-fungible-tokens" || return 1
+erdpy --verbose contract build "contracts/examples/ping-pong-egld" || return 1
+erdpy --verbose contract build "contracts/examples/simple-erc20" || return 1
+erdpy --verbose contract build "contracts/feature-tests/abi-tester" || return 1
+erdpy --verbose contract build "contracts/feature-tests/basic-features" || return 1
+erdpy --verbose contract build "contracts/feature-tests/async/async-alice" || return 1
+erdpy --verbose contract build "contracts/feature-tests/async/async-bob" || return 1
+erdpy --verbose contract build "contracts/feature-tests/use-module" || return 1
 
 ### CREATE ALL ABIs ###
