+++ elrond-wasm-derive/src/contract_impl.rs
@@ -96,10 +96,9 @@ pub fn contract_implementation(
 		#[allow(non_snake_case)]
 		pub mod endpoints {
 		  use super::*;
-		  use elrond_wasm_node::*;
 
-		  fn new_arwen_instance() -> #contract_impl_ident<ArwenApiImpl, ArwenBigInt, ArwenBigUint> {
-			let api = ArwenApiImpl{};
+		  fn new_arwen_instance() -> #contract_impl_ident<elrond_wasm_node::ArwenApiImpl, elrond_wasm_node::api::ArwenBigInt, elrond_wasm_node::api::ArwenBigUint> {
+			let api = elrond_wasm_node::ArwenApiImpl{};
 			#contract_impl_ident::new(api)
 		  }
