+++ elrond-wasm-node/src/api/error_api_node.rs
@@ -1,10 +1,10 @@
-use crate::ext_error;
+use crate::error_hook;
 use crate::ArwenApiImpl;
 use elrond_wasm::api::ErrorApi;
 
 impl ErrorApi for ArwenApiImpl {
 	#[inline]
 	fn signal_error(&self, message: &[u8]) -> ! {
-		ext_error::signal_error(message)
+		error_hook::signal_error(message)
 	}
 }
