+++ elrond-wasm-node/src/api/mod.rs
@@ -1,3 +1,6 @@
+mod big_int_api_node;
+mod big_uint_api_node;
+mod contract_api_node;
 mod crypto_api_node;
 mod endpoint_arg_api_node;
 mod endpoint_finish_api_node;
@@ -5,6 +8,9 @@ mod error_api_node;
 mod log_api_node;
 mod storage_api_node;
 
+pub use big_int_api_node::*;
+pub use big_uint_api_node::*;
+pub use contract_api_node::*;
 pub use crypto_api_node::*;
 pub use endpoint_arg_api_node::*;
 pub use endpoint_finish_api_node::*;
