+++ elrond-wasm-node/src/error_hook.rs
